+++ R/data.R
@@ -1,4 +1,4 @@
 #' Simulation conditions used in our study
 #' @format A data frame with 144 rows and 8 columns.
-#' See [check_conditions] for the columns description.
+#' See [prepare_conditions] for the columns description.
 "conditions"